backend/gateway/requirements.txt
fastapi
uvicorn
httpx
python-multipart
python-dotenv
pydantic
requests

